+++ src/Form/Flow/Type/FinishFlowType.php
@@ -4,12 +4,12 @@ namespace Yceruto\FormFlowBundle\Form\Flow\Type;
 
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Yceruto\FormFlowBundle\Form\Flow\AbstractFlowButtonType;
-use Yceruto\FormFlowBundle\Form\Flow\FlowButtonInterface;
-use Yceruto\FormFlowBundle\Form\Flow\FlowCursor;
+use Yceruto\FormFlowBundle\Form\Flow\AbstractButtonFlowType;
+use Yceruto\FormFlowBundle\Form\Flow\ButtonFlowInterface;
+use Yceruto\FormFlowBundle\Form\Flow\FormFlowCursor;
 use Yceruto\FormFlowBundle\Form\Flow\FormFlowInterface;
 
-class FlowFinishType extends AbstractFlowButtonType
+class FinishFlowType extends AbstractButtonFlowType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
@@ -19,8 +19,8 @@ class FlowFinishType extends AbstractFlowButtonType
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'handler' => fn (mixed $data, FlowButtonInterface $button, FormFlowInterface $flow) => $flow->reset(),
-            'include_if' => fn (FlowCursor $cursor): bool => $cursor->isLastStep(),
+            'handler' => fn (mixed $data, ButtonFlowInterface $button, FormFlowInterface $flow) => $flow->reset(),
+            'include_if' => fn (FormFlowCursor $cursor): bool => $cursor->isLastStep(),
         ]);
     }
 }