+++ tests/Fixtures/Flow/UserSignUpNavigatorType.php
@@ -5,14 +5,14 @@ namespace Yceruto\FormFlowBundle\Tests\Fixtures\Flow;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Yceruto\FormFlowBundle\Form\Flow\Type\FlowNavigatorType;
-use Yceruto\FormFlowBundle\Form\Flow\Type\FlowNextType;
+use Yceruto\FormFlowBundle\Form\Flow\Type\NavigatorFlowType;
+use Yceruto\FormFlowBundle\Form\Flow\Type\NextFlowType;
 
 class UserSignUpNavigatorType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        $builder->add('skip', FlowNextType::class, [
+        $builder->add('skip', NextFlowType::class, [
             'clear_submission' => true,
             'include_if' => ['professional'],
         ]);
@@ -27,6 +27,6 @@ class UserSignUpNavigatorType extends AbstractType
 
     public function getParent(): string
     {
-        return FlowNavigatorType::class;
+        return NavigatorFlowType::class;
     }
 }